About the project!
The project 'Tour Time' is completed by Aishwarya Singh and Simran Maurya. The project is planned and implemented by the two equally. Both the team members have written the code for the website. The website is suitable for mobile devices as well i.e. it is responsive. The project talks about the front-end of the website only. No back-end is implemented in this website and so no database is being used or created in this process.
The website contains hypertext(buttons and anchor elements), which will let the user to navigate to different pages of the websites. The website will give the user an experience of a Tours and Travels website. The website is user-friendly and solely created to enhance tourism in India.